chore: add babel import aliases
