+++ babel.config.js
@@ -6,6 +6,17 @@ module.exports = {
   plugins: [
     '@babel/plugin-transform-runtime',
     '@babel/plugin-proposal-class-properties',
-    '@babel/plugin-proposal-object-rest-spread'
+    '@babel/plugin-proposal-object-rest-spread',
+    [
+      'module-resolver',
+      {
+        root: ['./'],
+        alias: {
+          '@components': './src/components',
+          '@store': './src/store',
+          '@utils': './src/utils'
+        }
+      }
+    ]
   ]
 };
