+++ src/components/MTableEditRow/index.js
@@ -3,9 +3,9 @@ import TableCell from '@material-ui/core/TableCell';
 import TableRow from '@material-ui/core/TableRow';
 import Typography from '@material-ui/core/Typography';
 import PropTypes from 'prop-types';
-import { byString, setByString } from '../../utils';
-import * as CommonValues from '../../utils/common-values';
-import { validateInput } from '../../utils/validate';
+import { byString, setByString } from '@utils';
+import * as CommonValues from '@utils/common-values';
+import { validateInput } from '@utils/validate';
 
 function MTableEditRow(props) {
   const [state, setState] = useState(() => {
